# application/static/swagger.yaml
swagger: '2.0'
info:
  title: Mechanic Shop API
  description: REST API for managing customers, mechanics, service tickets, inventory, authentication, and advanced queries.
  version: '1.0.0'
host: 127.0.0.1:5000
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - Customers
      summary: Login a customer and return a JWT token.
      operationId: loginCustomer
      parameters:
        - in: body
          name: credentials
          required: true
          description: Email and password for the customer account.
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: Login succeeded.
          schema:
            $ref: '#/definitions/TokenResponse'
        '400':
          description: Validation error.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Invalid email or password.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /customers:
    post:
      tags:
        - Customers
      summary: Create a customer account.
      operationId: createCustomer
      parameters:
        - in: body
          name: customer
          required: true
          description: New customer account information.
          schema:
            $ref: '#/definitions/CustomerCreateRequest'
      responses:
        '201':
          description: Customer created.
          schema:
            $ref: '#/definitions/CustomerResponse'
        '400':
          description: Validation error or duplicate email.
          schema:
            $ref: '#/definitions/ErrorResponse'
    get:
      tags:
        - Customers
      summary: Get customers with pagination.
      operationId: getCustomers
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of customers to return.
          type: integer
          format: int32
          default: 10
        - name: offset
          in: query
          required: false
          description: Number of customers to skip before returning results.
          type: integer
          format: int32
          default: 0
      responses:
        '200':
          description: Customer list.
          schema:
            $ref: '#/definitions/CustomerListResponse'
        '400':
          description: Invalid pagination values.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: Get the authenticated customer's service tickets.
      operationId: getMyTickets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The authenticated customer's tickets.
          schema:
            $ref: '#/definitions/ServiceTicketListResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Get one customer by id.
      operationId: getCustomer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/parameters/CustomerId'
      responses:
        '200':
          description: Customer found.
          schema:
            $ref: '#/definitions/CustomerResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden for another customer's account.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - Customers
      summary: Update the authenticated customer's account.
      operationId: updateCustomer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/parameters/CustomerId'
        - in: body
          name: customer
          required: true
          description: Updated customer account information.
          schema:
            $ref: '#/definitions/CustomerCreateRequest'
      responses:
        '200':
          description: Customer updated.
          schema:
            $ref: '#/definitions/CustomerResponse'
        '400':
          description: Validation error.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden for another customer's account.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - Customers
      summary: Delete the authenticated customer's account.
      operationId: deleteCustomer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/parameters/CustomerId'
      responses:
        '200':
          description: Customer deleted.
          schema:
            $ref: '#/definitions/MessageResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Forbidden for another customer's account.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'
...
parameters:
  CustomerId:
    name: customer_id
    in: path
    required: true
    description: Customer id.
    type: integer
    format: int32
...
definitions:
  ErrorResponse:
    type: object
    additionalProperties: true
    properties:
      error:
        type: string
    example:
      error: Customer not found.
  MessageResponse:
    type: object
    properties:
      message:
        type: string
    example:
      message: Customer id 1 successfully deleted.
  TokenResponse:
    type: object
    properties:
      token:
        type: string
    example:
      token: eyJhbGciOi...
  LoginRequest:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        format: email
        example: customer@example.com
      password:
        type: string
        example: password123
    example:
      email: customer@example.com
      password: password123
  CustomerCreateRequest:
    type: object
    required:
      - name
      - email
      - phone
      - password
    properties:
      name:
        type: string
        example: Jamie Rivera
      email:
        type: string
        format: email
        example: jamie@example.com
      phone:
        type: string
        example: '555-123-4567'
      password:
        type: string
        example: password123
    example:
      name: Jamie Rivera
      email: jamie@example.com
      phone: '555-123-4567'
      password: password123
  CustomerResponse:
    type: object
    properties:
      id:
        type: integer
        format: int32
        example: 1
      name:
        type: string
        example: Jamie Rivera
      email:
        type: string
        format: email
        example: jamie@example.com
      phone:
        type: string
        example: '555-123-4567'
    example:
      id: 1
      name: Jamie Rivera
      email: jamie@example.com
      phone: '555-123-4567'
  PaginationResponse:
    type: object
    properties:
      limit:
        type: integer
        format: int32
        example: 10
      offset:
        type: integer
        format: int32
        example: 0
      count:
        type: integer
        format: int32
        example: 3
      total:
        type: integer
        format: int32
        example: 24
    example:
      limit: 10
      offset: 0
      count: 3
      total: 24
  CustomerListResponse:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/CustomerResponse'
      pagination:
        $ref: '#/definitions/PaginationResponse'
    example:
      data:
        - id: 1
          name: Jamie Rivera
          email: jamie@example.com
          phone: '555-123-4567'
      pagination:
        limit: 10
        offset: 0
        count: 1
        total: 1
...
  ServiceTicketResponse:
    type: object
    properties:
      id:
        type: integer
        format: int32
        example: 1
      VIN:
        type: string
        example: 1HGCM82633A004352
      service_date:
        type: string
        example: '2026-07-05'
      service_desc:
        type: string
        example: Brake replacement and inspection
      customer_id:
        type: integer
        format: int32
        example: 1
    example:
      id: 1
      VIN: 1HGCM82633A004352
      service_date: '2026-07-05'
      service_desc: Brake replacement and inspection
      customer_id: 1
  ServiceTicketListResponse:
    type: array
    items:
      $ref: '#/definitions/ServiceTicketResponse'
    example:
      - id: 1
        VIN: 1HGCM82633A004352
        service_date: '2026-07-05'
        service_desc: Brake replacement and inspection
        customer_id: 1
...